Finalise release which works with GHC-8.2 and base-4.10
* bump base dependency
* add new header files to cabal file
* bump version, update "tested-with", update stack resolver

+ fix a format string warning in Pack.c

+++ cbits/Pack.c
@@ -2028,7 +2028,7 @@ UnpackClosure (ClosureQ* q, HashTable* offsets,
     default:
         // invalid markers (not OFFSET, PLC, CLOSURE) are caught here
         errorBelch("unpackClosure: Found invalid marker %" FMT_Word ".\n",
-                   (long) **bufptrP);
+                   **bufptrP);
         return (StgClosure *) NULL;
     }
